request_api/requirements.txt
fastapi[all]
pandas
autogluon.tabular[lightgbm,catboost,xgboost,ray]